Code cleanup

+++ scripts/setup_emulator.py
@@ -8,27 +8,27 @@ def get_dolphin_dir():
     if platform.system() == "Windows":
         print("This is a Windows system")
         return os.path.join(os.environ["APPDATA"], "Dolphin Emulator")
-    
-    elif platform.system() == "Darwin":
-        print("This is a macOS system")
-        return os.path.join(os.path.expanduser("~"), "Library", "Application Support", "Dolphin")
-    
     else:
-        print("This is a Linux system")
-        return os.path.expanduser("~/.local/share/dolphin-emu")
-    
-SAVE_DST = os.path.join(get_dolphin_dir(), "Wii", "title",
-                        "00010004", "524d4350", "data", "rksys.dat")
+        raise RuntimeError("Unsupported OS for automatic Dolphin setup.")
 
-os.makedirs("emulator", exist_ok=True)
-print("Downloading Felk's Dolphin...")
-urllib.request.urlretrieve(RELEASE_URL, "emulator/dolphin.7z")
-print("Extracting...")
-with py7zr.SevenZipFile("emulator/dolphin.7z", mode='r') as z:
-    z.extractall("emulator/")
-os.remove("emulator/dolphin.7z")
-print("Done. Dolphin is at emulator/")
-print("Installing MKW save file...")
-os.makedirs(os.path.dirname(SAVE_DST), exist_ok=True)
-shutil.copy(os.path.join("assets", "rksys.dat"), SAVE_DST)
-print("Done, Dolphin is ready!.")
+if platform.system() == "Windows":
+    SAVE_DST = os.path.join(get_dolphin_dir(), "Wii", "title",
+                        "00010004", "524d4350", "data", "rksys.dat")
+    os.makedirs("emulator", exist_ok=True)
+    print("Downloading Felk's Dolphin...")
+    urllib.request.urlretrieve(RELEASE_URL, "emulator/dolphin.7z")
+    print("Extracting...")
+    with py7zr.SevenZipFile("emulator/dolphin.7z", mode='r') as z:
+        z.extractall("emulator/")
+    os.remove("emulator/dolphin.7z")
+    print("Done. Dolphin is at emulator/")
+    print("Installing MKW save file...")
+    os.makedirs(os.path.dirname(SAVE_DST), exist_ok=True)
+    shutil.copy(os.path.join("assets", "rksys.dat"), SAVE_DST)
+    print("Done, Dolphin is ready!.")
+
+elif platform.system() == "Darwin":
+    print("There is no macOS build of Felk's Dolphin. Please build it yourself from the source code.")
+
+else:
+    print("There is no Linux build of Felk's Dolphin. Please build it yourself from the source code.")
